+++ src/fuzzer.rs
@@ -1,16 +1,11 @@
-use core::marker::PhantomData;
 #[cfg(windows)]
 use std::ptr::write_volatile;
 use std::{
-    borrow::Cow,
     fs::create_dir_all,
+    marker::PhantomData,
     ops::DerefMut,
     path::PathBuf,
-    sync::{
-        atomic::{AtomicBool, Ordering},
-        Arc, Mutex,
-    },
-    time::{Duration, Instant},
+    sync::{Arc, Mutex},
 };
 
 use anyhow::{Context, Result};
@@ -19,25 +14,20 @@ use libafl::Fuzzer; // This may be marked unused, but will make the compiler giv
 use libafl::{
     corpus::{Corpus, OnDiskCorpus},
     events::{Event, EventFirer, SimpleEventManager},
-    executors::{
-        hooks::inprocess::inprocess_get_event_manager, inprocess::InProcessExecutor, Executor,
-        ExitKind, HasObservers,
-    },
+    executors::{Executor, ExitKind, HasObservers},
     feedback_or,
     feedbacks::{
         CrashFeedback, DifferentIsNovel, Feedback, MapFeedback, MaxMapFeedback, MaxReducer,
         TimeFeedback,
     },
     fuzzer::StdFuzzer,
-    inputs::BytesInput,
-    monitors::{AggregatorOps, UserStats, UserStatsValue},
     mutators::StdScheduledMutator,
     observers::{CanTrack, ExplicitTracking, MultiMapObserver, StdMapObserver, TimeObserver},
     schedulers::{
         powersched::PowerSchedule, IndexesLenTimeMinimizerScheduler, PowerQueueScheduler,
     },
     stages::{CalibrationStage, StdPowerMutationalStage},
-    state::{HasCorpus, HasExecutions, NopState},
+    state::{HasCorpus, HasExecutions},
     ExecuteInputResult, ExecutionProcessor, HasNamedMetadata,
 };
 use libafl_bolts::{
@@ -46,22 +36,16 @@ use libafl_bolts::{
     rands::StdRand,
     tuples::{tuple_list, MatchName},
 };
-use log::{debug, error, info};
+use log::{error, info};
 use openapiv3::OpenAPI;
 
 use crate::{
-    configuration::{Configuration, CrashCriterion},
+    configuration::Configuration,
     coverage_clients::{endpoint::EndpointCoverageClient, CoverageClient},
+    executor::SequenceExecutor,
     input::OpenApiInput,
     monitors::CoverageMonitor,
-    openapi::{
-        build_request::build_request_from_input,
-        curl_request::CurlRequest,
-        validate_response::{validate_response, Response},
-    },
     openapi_mutator::havoc_mutations_openapi,
-    parameter_feedback::ParameterFeedback,
-    reporting::Reporting,
     state::OpenApiFuzzerState,
 };
 
@@ -169,130 +153,14 @@ pub fn fuzz() -> Result<()> {
     let power = StdPowerMutationalStage::new(mutator_openapi);
     let mut stages = tuple_list!(calibration, power);
 
-    let (authentication, cookie_store, client) = crate::build_http_client()?;
-
-    let reporter = crate::reporting::sqlite::get_reporter(config)?;
-
-    // Keep track of the number of inputs
-    let mut inputs_tested = 0;
-    // Logging the number of executed requests
-    let mut stats = LoggingStats::new();
-
-    // The closure that we want to fuzz
-    let mut harness = |inputs: &OpenApiInput| {
-        let mut exit_kind = ExitKind::Ok;
-        inputs_tested += 1;
-
-        let mut parameter_feedback = ParameterFeedback::new(inputs.0.len());
-        log::debug!("Sending {} requests", inputs.0.len());
-        'chain: for (request_index, request) in inputs.0.iter().enumerate() {
-            let mut request = request.clone();
-            log::trace!("OpenAPI request:\n{:#?}", request);
-            if let Err(error) = request.resolve_parameter_references(&parameter_feedback) {
-                debug!(
-                        "Cannot instantiate request: missing value for backreferenced parameter: {}. Maybe the earlier request crashed?",
-                        error
-                    );
-                break 'chain;
-            };
-            let request_builder =
-                match build_request_from_input(&client, &cookie_store, &api, &request) {
-                    None => continue,
-                    Some(r) => r.timeout(Duration::from_millis(config.request_timeout)),
-                };
-
-            let request_built = match request_builder.build() {
-                Ok(request) => request,
-                Err(err) => {
-                    // We don't expect errors to occur in the reqwest builder. If one occurs,
-                    // it's not the target's fault, so we don't set ExitKind::Crash or Timeout.
-                    error!("Error building request: {err}");
-                    break;
-                }
-            };
-
-            let curl_request = CurlRequest(&request_built, &authentication);
-            let reporter_request_id =
-                reporter.report_request(&request, &curl_request, inputs_tested);
-            let curl_request = curl_request.to_string();
-
-            match client.execute(request_built) {
-                Ok(response) => {
-                    stats.performed_requests += 1;
-                    let response: Response = response.into();
-
-                    endpoint_coverage_client.lock().unwrap().cover(
-                        request.method,
-                        request.path.clone(),
-                        response.status(),
-                        curl_request,
-                        response.text().unwrap_or_else(|_| {
-                            String::from("Unable to decode the response to UTF-8")
-                        }),
-                    );
-                    reporter.report_response(&response, reporter_request_id);
-                    log::trace!("Got response {}", response.status());
-
-                    if response.status() == 429 {
-                        log::warn!("HTTP status 429 'Too Many Requests' encountered!");
-                        log::warn!("Rate limiting is likely active on the program under test.");
-                        log::warn!("This hinders fuzz testing. Consider disabling it.");
-                    }
-
-                    if response.status().is_server_error() {
-                        exit_kind = ExitKind::Crash;
-                        log::debug!("OpenAPI-input resulted in server error response, ignoring rest of request chain.");
-                        break 'chain;
-                    } else {
-                        if config.crash_criterion == CrashCriterion::AllErrors {
-                            if let Err(validation_err) =
-                                validate_response(&api, &request, &response)
-                            {
-                                log::debug!("OpenAPI-input resulted in validation error: {validation_err}, ignoring rest of request chain.");
-                                exit_kind = ExitKind::Crash;
-                                break 'chain;
-                            }
-                        }
-                        if response.status().is_success() {
-                            parameter_feedback.process_response(request_index, response);
-                        }
-                    }
-                }
-                Err(e) => {
-                    reporter.report_response_error(&e.to_string(), reporter_request_id);
-                    error!("{}", e);
-                    exit_kind = ExitKind::Timeout;
-                    log::debug!(
-                        "OpenAPI-request resulted in timeout, ignoring rest of request chain."
-                    );
-                    break;
-                }
-            }
-            parameter_feedback.process_post_request(request_index, request);
-        }
-        update_coverage(
-            &mut code_coverage_client,
-            &mut endpoint_coverage_client,
-            &reporter,
-            &mut stats,
-            |s: String| info!("{}", s),
-        );
-
-        exit_kind
-    };
-
     // Create the executor for an in-process function with just one observer
-    let mut executor = InProcessExecutor::with_timeout(
-        &mut harness,
+    let mut executor = SequenceExecutor::new(
         collective_observer,
-        &mut fuzzer,
-        &mut state,
-        &mut mgr,
-        Duration::from_millis(0), // disable the timeout
-    )
-    .context("Failed to create the Executor")?;
-
-    let manual_interrupt = setup_interrupt()?;
+        &api,
+        config,
+        code_coverage_client,
+        endpoint_coverage_client.clone(),
+    )?;
 
     // Fire an event to print the initial corpus size
     let corpus_size = state.corpus().count();
@@ -328,17 +196,12 @@ pub fn fuzz() -> Result<()> {
     }
 
     log::debug!("Start fuzzing loop");
-    let maybe_timeout_secs = config.timeout.map(|t| Duration::from_secs(t.get()));
-    let starting_time = Instant::now();
-    // check for timeout if applicable
-    while maybe_timeout_secs
-        .map(|timeout| Instant::now() - starting_time < timeout)
-        .unwrap_or(true)
-    {
+    loop {
         match fuzzer.fuzz_one(&mut stages, &mut executor, &mut state, &mut mgr) {
             Ok(_) => (),
             Err(libafl_bolts::Error::ShuttingDown) => {
-                return Ok(());
+                log::info!("[Fuzzing campaign ended] Thanks for using WuppieFuzz!");
+                break;
             }
             Err(err) => {
                 return Err(err).context("Error in the fuzz loop");
@@ -356,17 +219,10 @@ pub fn fuzz() -> Result<()> {
         ) {
             error!("Err: failed to fire event{:?}", e)
         }
-        if manual_interrupt.load(Ordering::Relaxed) {
-            if let Err(e) = mgr.fire(&mut state, Event::Stop) {
-                error!("Err: failed to fire event{:?}", e);
-                break;
-            }
-        }
     }
 
     if let Some(report_path) = report_path {
-        endpoint_coverage_client.generate_coverage_report(&report_path);
-        code_coverage_client.generate_coverage_report(&report_path);
+        executor.generate_coverage_report(&report_path);
     }
 
     Ok(())
@@ -452,24 +308,6 @@ fn setup_line_coverage<'a>(
     ))
 }
 
-/// Installs the Ctrl-C interrupt handler
-fn setup_interrupt() -> Result<Arc<AtomicBool>, anyhow::Error> {
-    let manual_interrupt = Arc::new(AtomicBool::new(false));
-    {
-        let manual_interrupt = Arc::clone(&manual_interrupt);
-        ctrlc::set_handler(move || {
-            let second_time_pressed = manual_interrupt.swap(true, Ordering::Relaxed);
-            if second_time_pressed {
-                info!("Ctrl + c pressed, again - exiting forcefully!");
-                std::process::exit(0);
-            } else {
-                info!("Ctrl + c pressed, starting graceful shutdown.");
-            }
-        })?;
-    }
-    Ok(manual_interrupt)
-}
-
 /// Creates and returns the report path for this run. It is typically of the form
 /// `reports/2023-06-13T105302.602Z`, the filename being an ISO 8601 timestamp.
 fn generate_report_path() -> PathBuf {
@@ -480,112 +318,4 @@ fn generate_report_path() -> PathBuf {
     let report_path = PathBuf::from("reports").join(timestamp);
     create_dir_all(&report_path).expect("unable to make reports directory");
     report_path
-}
-
-/// How often to print a new log line
-const CLIENT_STATS_TIME_WINDOW_SECS: u64 = 5;
-
-#[derive(Clone, Copy)]
-struct LoggingStats {
-    performed_requests: u64,
-    last_window_time: Instant,
-    last_covered: u64,
-    last_endpoint_covered: u64,
-}
-
-impl LoggingStats {
-    fn new() -> Self {
-        Self {
-            performed_requests: 0,
-            last_window_time: Instant::now(),
-            last_covered: 0,
-            last_endpoint_covered: 0,
-        }
-    }
-}
-
-/// Updates the LibAFL event manager with coverage information fetched from the
-/// endpoint coverage monitor and any line coverage client. Also updates the given
-/// (MySqLite) reporter
-fn update_coverage<F: FnMut(String)>(
-    code_coverage_client: &mut Box<dyn CoverageClient>,
-    endpoint_coverage_client: &mut Arc<Mutex<EndpointCoverageClient>>,
-    reporter: &dyn Reporting<i64>,
-    stats: &mut LoggingStats,
-    _print_fn: F,
-) {
-    code_coverage_client.fetch_coverage(true);
-    let (covered, total) = code_coverage_client.max_coverage_ratio();
-    endpoint_coverage_client.fetch_coverage(true);
-    let (e_covered, e_total) = endpoint_coverage_client.max_coverage_ratio();
-
-    // This input is needed for event_manager.fire, but it doesn't seem to make
-    // a difference whether it is meaningful or not, and cloning the entire thing
-    // in the harness (because it's immutable there and we need it to be mutable)
-    // seems wasteful.
-    let mut state = NopState::new();
-
-    // Add own user stats
-    let cov_stats = UserStatsValue::Ratio(covered, total);
-    let end_cov_stats = UserStatsValue::Ratio(e_covered, e_total);
-
-    let req_stats = UserStatsValue::Number(stats.performed_requests);
-
-    let event_manager = unsafe {
-        inprocess_get_event_manager::<
-            SimpleEventManager<BytesInput, CoverageMonitor<F>, NopState<BytesInput>>,
-        >()
-        .expect("Can not load the event manager")
-    };
-
-    if covered != stats.last_covered {
-        stats.last_covered = covered;
-        // send the coverage stats to the event manager for use in the monitor
-        if let Err(e) = event_manager.fire(
-            &mut state,
-            Event::UpdateUserStats {
-                name: Cow::Borrowed("wuppiefuzz_code_coverage"),
-                value: UserStats::new(cov_stats, AggregatorOps::None),
-                phantom: PhantomData,
-            },
-        ) {
-            error!("Err: failed to fire event{:?}", e)
-        }
-    }
-
-    if e_covered != stats.last_endpoint_covered {
-        stats.last_endpoint_covered = e_covered;
-        // send the coverage stats to the event manager for use in the monitor
-        if let Err(e) = event_manager.fire(
-            &mut state,
-            Event::UpdateUserStats {
-                name: Cow::Borrowed("wuppiefuzz_endpoint_coverage"),
-                value: UserStats::new(end_cov_stats, AggregatorOps::None),
-                phantom: PhantomData,
-            },
-        ) {
-            error!("Err: failed to fire event{:?}", e)
-        }
-    }
-
-    let current_time = Instant::now();
-    let diff = current_time
-        .duration_since(stats.last_window_time)
-        .as_secs();
-    if diff > CLIENT_STATS_TIME_WINDOW_SECS {
-        stats.last_window_time = current_time;
-        // send the request stats to the event manager for use in the monitor
-        if let Err(e) = event_manager.fire(
-            &mut state,
-            Event::UpdateUserStats {
-                name: Cow::Borrowed("requests"),
-                value: UserStats::new(req_stats, AggregatorOps::None),
-                phantom: PhantomData,
-            },
-        ) {
-            error!("Err: failed to fire event{:?}", e)
-        }
-    }
-
-    reporter.report_coverage(covered, total, e_covered, e_total)
 }