+++ CHANGELOG.md
@@ -4,6 +4,9 @@
 
 ## Features
 
+- Adds custom executor enabling proper timeout kill and removing unsafe code in
+  [#49](https://github.com/TNO-S3/WuppieFuzz/pull/49)
+
 ## Fixes
 
 - Updated [LibAFL](https://github.com/AFLplusplus/LibAFL) to 0.15.0 in